Last commit is
commit 57fd524448827e6e030725ce869e46f2d5b05fc2
Author: xieguo <xieguo.xg@alibaba-inc.com>
Date:   Sun Feb 15 11:15:17 2015 +0800

    Remove unused code to avoid compile failure.

Conflicts:
	pom.xml
	src/com/taobao/tao/TaobaoApplication.java

+++ src/com/taobao/tao/atlaswrapper/OptDexProcess.java
@@ -31,34 +31,55 @@ public class OptDexProcess {
 		mIsInited = true;
 	}	
 	
-	public synchronized  void processPackages(){
+	public synchronized  void processPackages(boolean onlyAuto, boolean force){
 		// Never process once not initialized yet to avoid null exception
 		if (!mIsInited){
 			Log.e(TAG, "Bundle Installer not initialized yet, process abort!");
 			return;
-		} else if (mIsProcessed){
+		} else if (mIsProcessed && !force){
 			Log.i(TAG, "Bundle install already executed, just return");
 			return;
 		}
 		
-    	// 完成非delayed bundle的dexopt		
-        long start = System.currentTimeMillis();
-		processPakagesNotDelayed();
-        Log.d(TAG, "Install bundles not delayed cost time = " + (System.currentTimeMillis() - start) + " ms");
-        
-        // Send out BUNDLES_INSTALLED intent to inform welcome can start the HOME activity now.
-        Utils.saveAtlasInfoBySharedPreferences(mApplication);
-        System.setProperty("BUNDLES_INSTALLED", "true");
-        notifyBundleInstalled();
-        
-		// 完成delayed bundle的dexopt
-        start = System.currentTimeMillis();
-    	OptDexProcess mOptDexProcess = OptDexProcess.getInstance();
-    	mOptDexProcess.processPakcagesDelayed();
-    	
-        Log.d(TAG, "Install delayed bundles cost time = " + (System.currentTimeMillis() - start) + " ms");    	
-        
-    	mIsProcessed = true;    	
+		if (onlyAuto) {
+	    	// 完成auto start bundle的dexopt		
+	        long start = System.currentTimeMillis();
+			processPakcagesAutoStarted();
+			if (!force){
+				NotifyBundleInstalled();
+			}
+	        Log.d(TAG, "dexopt auto start bundles cost time = " + (System.currentTimeMillis() - start) + " ms");
+		} else {
+	    	// 完成非delayed bundle的dexopt		
+	        long start = System.currentTimeMillis();
+			processPakagesNotDelayed();
+	        Log.d(TAG, "dexopt bundles not delayed cost time = " + (System.currentTimeMillis() - start) + " ms");
+			if (!force){
+				NotifyBundleInstalled();
+			}
+			// 完成delayed bundle的dexopt
+	        start = System.currentTimeMillis();
+	    	OptDexProcess mOptDexProcess = OptDexProcess.getInstance();
+	    	mOptDexProcess.processPakcagesDelayed();
+	        Log.d(TAG, "dexopt delayed bundles cost time = " + (System.currentTimeMillis() - start) + " ms");    	
+		}
+		
+        /*
+         *  Mark process flag as true to avoid bundle installation executed twice.
+         *  Dont's set the flag when force since below logic:
+         *  replaced receiver excecute-->TaobaoApplication onCreate()'s thread execution
+         *  NotifyBundleInstalled() wouldn't be execute.
+         */		
+		if (!force){
+			mIsProcessed = true;    	
+		}
+	}
+
+	private void NotifyBundleInstalled() {
+		// Send out BUNDLES_INSTALLED intent to inform welcome can start the HOME activity now.
+		Utils.saveAtlasInfoBySharedPreferences(mApplication);
+		System.setProperty("BUNDLES_INSTALLED", "true");
+		mApplication.sendBroadcast(new Intent("com.taobao.taobao.action.BUNDLES_INSTALLED"));
 	}
 	
 	/*
@@ -83,14 +104,29 @@ public class OptDexProcess {
         }
         
 	}	
-	
-    private void notifyBundleInstalled(){
-    	mApplication.sendBroadcast(new Intent("com.taobao.taobao.action.BUNDLES_INSTALLED"));
-    }
     
 	private void processPakcagesDelayed(){
 		
 		for (String pkg : Utils.DELAYED_PACKAGES) {
+			Bundle bundle = Atlas.getInstance().getBundle(pkg);
+			if (bundle != null) {
+				try {
+					((BundleImpl) bundle).optDexFile();
+                    Atlas.getInstance().enableComponent(bundle.getLocation());					
+				} catch (Exception e) {
+                	if (e instanceof BundleArchiveRevision.DexLoadException){
+                		throw (RuntimeException)e;
+                	}
+                    Log.e(TAG, "Error while dexopt >>>", e);
+				}
+			}
+		}
+		
+	}	
+	
+	private void processPakcagesAutoStarted(){
+		
+		for (String pkg : Utils.AUTOSTART_PACKAGES) {
 			Bundle bundle = Atlas.getInstance().getBundle(pkg);
 			if (bundle != null) {
 				try {
